backend/requirements.txt
Django==5.2.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
boto3==1.34.34
django-storages==1.14.2
Pillow==10.2.0
python-dateutil==2.8.2
dj-database-url==2.1.0 